migrate environment context to world state
