+++ codex-rs/core/src/context/world_state/environment.rs
@@ -1,14 +1,13 @@
 use super::WorldStateSection;
+use super::environment_support::FileSystemContext;
+use super::environment_support::NetworkContext;
+use super::environment_support::push_xml_escaped_text;
 use crate::context::ContextualUserFragment;
-use crate::context::environment_context::EnvironmentContext;
-use crate::context::environment_context::EnvironmentContextEnvironment;
-use crate::context::environment_context::EnvironmentContextEnvironments;
-use crate::context::environment_context::FileSystemContext;
-use crate::context::environment_context::NetworkContext;
-use crate::context::environment_context::push_xml_escaped_text;
 use crate::session::turn_context::TurnContext;
-use crate::shell::Shell;
 use codex_protocol::models::ResponseItem;
+use codex_protocol::protocol::TurnContextItem;
+use codex_protocol::protocol::TurnContextNetworkItem;
+use codex_utils_absolute_path::AbsolutePathBuf;
 use codex_utils_path_uri::PathUri;
 use serde::Deserialize;
 use serde::Serialize;
@@ -40,14 +39,35 @@ impl PartialEq for EnvironmentsState {
 impl Eq for EnvironmentsState {}
 
 impl EnvironmentsState {
-    pub(crate) fn from_turn_context(
-        turn_context: &TurnContext,
-        default_shell: &Shell,
-        subagents: String,
-    ) -> Self {
-        let context = EnvironmentContext::from_turn_context(turn_context, default_shell)
-            .with_subagents(subagents);
-        let mut state = Self::from_environment_context(&context);
+    pub(crate) fn from_turn_context(turn_context: &TurnContext) -> Self {
+        let mut state = Self {
+            environments: turn_context
+                .environments
+                .turn_environments
+                .iter()
+                .map(|environment| {
+                    (
+                        environment.environment_id.clone(),
+                        EnvironmentState {
+                            cwd: environment.cwd().clone(),
+                            status: Some(EnvironmentStatus::Available),
+                            shell: environment
+                                .shell
+                                .as_ref()
+                                .map(|shell| shell.name().to_string()),
+                        },
+                    )
+                })
+                .collect(),
+            current_date: turn_context.current_date.clone(),
+            timezone: turn_context.timezone.clone(),
+            network: network_from_turn_context(turn_context),
+            filesystem: Some(FileSystemContext::from_permission_profile(
+                &turn_context.permission_profile,
+                &turn_context.config.effective_workspace_roots(),
+            )),
+            subagents: None,
+        };
         for environment in &turn_context.environments.starting {
             state
                 .environments
@@ -61,26 +81,55 @@ impl EnvironmentsState {
         state
     }
 
-    pub(crate) fn from_environment_context(context: &EnvironmentContext) -> Self {
-        let mut environments = BTreeMap::new();
-        match &context.environments {
-            EnvironmentContextEnvironments::None => {}
-            EnvironmentContextEnvironments::Single(environment) => {
-                insert_environment(&mut environments, environment);
-            }
-            EnvironmentContextEnvironments::Multiple(current) => {
-                for environment in current {
-                    insert_environment(&mut environments, environment);
-                }
-            }
-        }
+    pub(crate) fn from_turn_context_item(turn_context_item: &TurnContextItem) -> Self {
         Self {
-            environments,
-            current_date: context.current_date.clone(),
-            timezone: context.timezone.clone(),
-            network: context.network.clone(),
-            filesystem: context.filesystem.clone(),
-            subagents: context.subagents.clone(),
+            environments: [(
+                String::new(),
+                EnvironmentState {
+                    cwd: PathUri::from_abs_path(&turn_context_item.cwd),
+                    status: Some(EnvironmentStatus::Available),
+                    shell: None,
+                },
+            )]
+            .into_iter()
+            .collect(),
+            current_date: turn_context_item.current_date.clone(),
+            timezone: turn_context_item.timezone.clone(),
+            network: network_from_turn_context_item(turn_context_item),
+            filesystem: Some(FileSystemContext::from_permission_profile(
+                &turn_context_item.permission_profile(),
+                &workspace_roots_from_turn_context_item(turn_context_item),
+            )),
+            subagents: None,
+        }
+    }
+
+    pub(crate) fn with_subagents(mut self, subagents: String) -> Self {
+        if !subagents.is_empty() {
+            self.subagents = Some(subagents);
+        }
+        self
+    }
+
+    pub(crate) fn render_diff(&self, previous: &Self) -> Option<ResponseItem> {
+        WorldStateSection::render_diff(self, previous)
+    }
+
+    fn rendered_full(&self) -> RenderedEnvironments {
+        RenderedEnvironments {
+            updates: self
+                .environments
+                .iter()
+                .map(|(id, environment)| {
+                    (id.clone(), EnvironmentUpdate::Current(environment.clone()))
+                })
+                .collect(),
+            legacy_single: is_legacy_single(&self.environments),
+            current_date: self.current_date.clone(),
+            timezone: self.timezone.clone(),
+            network: self.network.clone(),
+            filesystem: self.filesystem.clone(),
+            subagents: self.subagents.clone(),
         }
     }
 }
@@ -89,22 +138,32 @@ impl WorldStateSection for EnvironmentsState {
     const NAME: &'static str = "environments";
 
     fn render_diff(&self, previous: &Self) -> Option<ResponseItem> {
+        let legacy_single =
+            is_legacy_single(&self.environments) && previous.environments.len() <= 1;
+        if legacy_single
+            && self.environments.values().next() == previous.environments.values().next()
+        {
+            return None;
+        }
         let mut updates = self
             .environments
             .iter()
             .filter(|(id, environment)| previous.environments.get(*id) != Some(*environment))
             .map(|(id, environment)| (id.clone(), EnvironmentUpdate::Current(environment.clone())))
             .collect::<BTreeMap<_, _>>();
-        updates.extend(
-            previous
-                .environments
-                .keys()
-                .filter(|id| !self.environments.contains_key(*id))
-                .map(|id| (id.clone(), EnvironmentUpdate::Unavailable)),
-        );
+        if !legacy_single {
+            updates.extend(
+                previous
+                    .environments
+                    .keys()
+                    .filter(|id| !id.is_empty() && !self.environments.contains_key(*id))
+                    .map(|id| (id.clone(), EnvironmentUpdate::Unavailable)),
+            );
+        }
         (!updates.is_empty()).then(|| {
             ContextualUserFragment::into(RenderedEnvironments {
                 updates,
+                legacy_single,
                 current_date: self.current_date.clone(),
                 timezone: self.timezone.clone(),
                 network: self.network.clone(),
@@ -115,8 +174,27 @@ impl WorldStateSection for EnvironmentsState {
     }
 }
 
+impl ContextualUserFragment for EnvironmentsState {
+    fn role(&self) -> &'static str {
+        "user"
+    }
+
+    fn markers(&self) -> (&'static str, &'static str) {
+        Self::type_markers()
+    }
+
+    fn type_markers() -> (&'static str, &'static str) {
+        environment_context_markers()
+    }
+
+    fn body(&self) -> String {
+        self.rendered_full().body()
+    }
+}
+
 struct RenderedEnvironments {
     updates: BTreeMap<String, EnvironmentUpdate>,
+    legacy_single: bool,
     current_date: Option<String>,
     timezone: Option<String>,
     network: Option<NetworkContext>,
@@ -139,15 +217,16 @@ impl ContextualUserFragment for RenderedEnvironments {
     }
 
     fn type_markers() -> (&'static str, &'static str) {
-        (
-            codex_protocol::protocol::ENVIRONMENT_CONTEXT_OPEN_TAG,
-            codex_protocol::protocol::ENVIRONMENT_CONTEXT_CLOSE_TAG,
-        )
+        environment_context_markers()
     }
 
     fn body(&self) -> String {
         let mut rendered = "\n".to_string();
-        if !self.updates.is_empty() {
+        if self.legacy_single {
+            if let Some(EnvironmentUpdate::Current(environment)) = self.updates.values().next() {
+                push_environment_values(&mut rendered, environment, "  ");
+            }
+        } else if !self.updates.is_empty() {
             rendered.push_str("  <environments>\n");
             for (id, update) in &self.updates {
                 match update {
@@ -160,18 +239,9 @@ impl ContextualUserFragment for RenderedEnvironments {
                             rendered.push_str(status.as_str());
                             rendered.push('"');
                         }
-                        rendered.push_str(">\n      <cwd>");
-                        push_xml_escaped_text(
-                            &mut rendered,
-                            &environment.cwd.inferred_native_path_string(),
-                        );
-                        rendered.push_str("</cwd>");
-                        if let Some(shell) = &environment.shell {
-                            rendered.push_str("\n      <shell>");
-                            push_xml_escaped_text(&mut rendered, shell);
-                            rendered.push_str("</shell>");
-                        }
-                        rendered.push_str("\n    </environment>\n");
+                        rendered.push_str(">\n");
+                        push_environment_values(&mut rendered, environment, "      ");
+                        rendered.push_str("    </environment>\n");
                     }
                     EnvironmentUpdate::Unavailable => {
                         rendered.push_str("    <environment id=\"");
@@ -207,18 +277,17 @@ impl ContextualUserFragment for RenderedEnvironments {
     }
 }
 
-fn insert_environment(
-    environments: &mut BTreeMap<String, EnvironmentState>,
-    environment: &EnvironmentContextEnvironment,
-) {
-    environments.insert(
-        environment.id.clone(),
-        EnvironmentState {
-            cwd: environment.cwd.clone(),
-            status: Some(EnvironmentStatus::Available),
-            shell: Some(environment.shell.clone()),
-        },
-    );
+fn push_environment_values(rendered: &mut String, environment: &EnvironmentState, indent: &str) {
+    rendered.push_str(indent);
+    rendered.push_str("<cwd>");
+    push_xml_escaped_text(rendered, &environment.cwd.inferred_native_path_string());
+    rendered.push_str("</cwd>\n");
+    if let Some(shell) = &environment.shell {
+        rendered.push_str(indent);
+        rendered.push_str("<shell>");
+        push_xml_escaped_text(rendered, shell);
+        rendered.push_str("</shell>\n");
+    }
 }
 
 fn push_optional_element(rendered: &mut String, name: &str, value: Option<&str>) {
@@ -266,6 +335,71 @@ impl EnvironmentStatus {
     }
 }
 
+fn is_legacy_single(environments: &BTreeMap<String, EnvironmentState>) -> bool {
+    environments.len() == 1
+        && matches!(
+            environments
+                .values()
+                .next()
+                .and_then(|environment| environment.status),
+            Some(EnvironmentStatus::Available)
+        )
+}
+
+fn environment_context_markers() -> (&'static str, &'static str) {
+    (
+        codex_protocol::protocol::ENVIRONMENT_CONTEXT_OPEN_TAG,
+        codex_protocol::protocol::ENVIRONMENT_CONTEXT_CLOSE_TAG,
+    )
+}
+
+fn network_from_turn_context(turn_context: &TurnContext) -> Option<NetworkContext> {
+    let network = turn_context
+        .config
+        .config_layer_stack
+        .requirements()
+        .network
+        .as_ref()?;
+
+    Some(NetworkContext::new(
+        network
+            .domains
+            .as_ref()
+            .and_then(codex_config::NetworkDomainPermissionsToml::allowed_domains)
+            .unwrap_or_default(),
+        network
+            .domains
+            .as_ref()
+            .and_then(codex_config::NetworkDomainPermissionsToml::denied_domains)
+            .unwrap_or_default(),
+    ))
+}
+
+fn network_from_turn_context_item(turn_context_item: &TurnContextItem) -> Option<NetworkContext> {
+    let TurnContextNetworkItem {
+        allowed_domains,
+        denied_domains,
+    } = turn_context_item.network.as_ref()?;
+    Some(NetworkContext::new(
+        allowed_domains.clone(),
+        denied_domains.clone(),
+    ))
+}
+
+fn workspace_roots_from_turn_context_item(
+    turn_context_item: &TurnContextItem,
+) -> Vec<AbsolutePathBuf> {
+    if let Some(workspace_roots) = turn_context_item.workspace_roots.as_ref() {
+        return workspace_roots.clone();
+    }
+
+    vec![turn_context_item.cwd.clone()]
+}
+
 #[cfg(test)]
 #[path = "environment_tests.rs"]
 mod tests;
+
+#[cfg(test)]
+#[path = "environment_render_tests.rs"]
+mod render_tests;
